monitoring-servers-side/monitoring-services/requirements.txt
fastapi==0.95.2
gunicorn==20.1.0
loguru==0.7.2
pydantic==1.10.8
PyYAML==5.3.1
uvicorn==0.22.0
